04-Python-Unit-Testing/Example-1/pyproject.toml
[project]
name = "html-color"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.10.1",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pyhamcrest>=2.1.0",
    "pytest>=8.3.5",
    "ruff>=0.11.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
# Thank you... <https://github.com/astral-sh/uv/issues/7260#issuecomment-2341586054>
pythonpath = ["."]

[tool.coverage.run]
omit = [
    "/usr/*",
    ".tox/*",
    "tests/*"
]

[tool.coverage.report]
exclude_also = [
    "def main"
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 80

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.lint]
select = ["D101", "E", "W", "F", "B", "Q", "I001"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]



